fix: 修改類型

+++ const/worker/worker.go
@@ -11,14 +11,14 @@ var (
 )
 
 type Cookie struct {
-	Name     string `json:"name"`
-	Value    string `json:"value"`
-	Path     string `json:"path,omitempty"`
-	Domain   string `json:"domain,omitempty"`
-	Expires  int64  `json:"expirationDate,omitempty"`
-	Secure   bool   `json:"secure,omitempty"`
-	HttpOnly bool   `json:"httpOnly,omitempty"`
-	SameSite string `json:"sameSite,omitempty"`
+	Name     string  `json:"name"`
+	Value    string  `json:"value"`
+	Path     string  `json:"path,omitempty"`
+	Domain   string  `json:"domain,omitempty"`
+	Expires  float64 `json:"expirationDate,omitempty"`
+	Secure   bool    `json:"secure,omitempty"`
+	HttpOnly bool    `json:"httpOnly,omitempty"`
+	SameSite string  `json:"sameSite,omitempty"`
 }
 
 type AuctionItemInfo struct {
